Add opt-in containerd support
New labels:
container_runtime, containerd or fallback to host-docker
containerd_version, taken from https://github.com/containerd/containerd/releases
containerd_tarball_url, eg https://storage.googleapis.com/cri-containerd-release/cri-containerd-1.2.4.linux-amd64.tar.gz
containerd_tarball_sha256, sha256 of the above tarball

story: 2007317
task: 38823

(cherry picked from commit de21e0431aa4785184d43b79525f610c9f07a4eb)

+++ magnum/drivers/common/templates/kubernetes/fragments/enable-services-minion.sh
@@ -8,13 +8,20 @@ ssh_cmd="ssh -F /srv/magnum/.ssh/config root@localhost"
 # so we need to stop it first and delete the docker0 bridge (which will
 # be re-created using the flannel-provided subnet).
 echo "stopping docker"
-$ssh_cmd systemctl stop docker
+if [ ${CONTAINER_RUNTIME} != "containerd"  ] ; then
+    $ssh_cmd systemctl stop docker
+fi
 
 # make sure we pick up any modified unit files
 $ssh_cmd systemctl daemon-reload
 
+if [ ${CONTAINER_RUNTIME} = "containerd"  ] ; then
+    container_runtime_service="containerd"
+else
+    container_runtime_service="docker"
+fi
 for action in enable restart; do
-    for service in docker kubelet kube-proxy; do
+    for service in ${container_runtime_service} kubelet kube-proxy; do
         echo "$action service $service"
         $ssh_cmd systemctl $action $service
     done
